lib/igt_fb: Add support for NV12 format through conversion
For NV12 a format conversion is needed. Because YUV formats are not
fully defined with just a fourcc, I've chosen BT.709 limited range.
This puts the pixel center of the CbCr components between the top
left Y and bottom left Y:

Y   Y   Y   Y
UV      UV
Y   Y   Y   Y

Some work is put into optimising the conversion routines in order to
make it fast enough. Before converting nv12 to rgb24, it is copied to
a temporary buffer to take advantage of memory caching. This is
approximately 20x faster than directly reading the BO.

When testing on my KBL with a 1080p buffer, it takes approximately
.1s to convert either way, this is fast enough not to bother optimising
even further for me.

Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>