tests/kms_concurrent: Use maximum number of planes that display/platform combination supports
There was an error in

commit 0ab05a51a059 ("tests/kms_concurrent: Test for maximum number of planes")

that calculates the maximum number of supported planes. The patch proposes a fix
by testing first, if a successful atomic commit is able to do with the selected
number of planes. In case of a failure, the number of planes is reduced by one
assuming that the failure was caused by bandwidth limit. The test loops at least
at many iterations as there are planes defined by the platform.

Signed-off-by: Mika Kahola <mika.kahola@intel.com>