igt: Make igt_simple_main*() look like normal function definition
Coccinelle chokes on igt_simple_main*() because it looks like
a function definition without a type. Make it look like
a normal thing by moving the type out of the macro.

The macro still needs the __real_main forward declaration
so we can't actually expose the type from the main()
definition, but we can add a forward declaration for main()
ahead of the __real_main declaration, and if we leave out
the type there it must be placed into the place where
igt_simple_main*() is used.

The bulk conversion was done with sed:
$ find -type f -name \*.[ch] -exec sed -i 's/^igt_simple_main/int igt_simple_main/g' {} +

Acked-by: Andrzej Hajda <andrzej.hajda@intel.com>
Acked-by: Kamil Konieczny <kamil.konieczny@linux.intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>