lib/kms: Turn the igt_plane*prop*() macros into functions
Use functions instead of macros for the plane property stuff.
Macros are generally not type safe, hideous to look at, and
a pain to modify. So using macros when you don't have to
is simply a bad idea.

With the type safe functions we now see that tests/kms_cursor_legacy
was using the wrong macros to set cursor properties. But I guess
in this case the implementation of the wrong macros was close
enough that they actually worked correctly.

Acked-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>