tests/intel/xe_oa: Drop oa-unit-exclusive-stream-exec-q
In the kernel we have restricted exec_queue's to be specified only for the
OAG unit. Therefore the "oa-unit-exclusive-stream-exec-q" IGT test needs to
be modified. However, the main purpose of the "oa-unit-exclusive-stream"
tests is to ensure that only a single OA stream can be opened on an OA
unit. For this an exec_queue is not needed, just the sampling mode is
sufficient. Therefore, discontinue oa-unit-exclusive-stream-exec-q test. It
can be added back later if it is determined that it is really needed.

Signed-off-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Reviewed-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>