lib/kms: Get rid of the 'p' plane index variable
The 'p' thing used to track the next free plane slot is annoying.
Get rid of it by just initializing the plane indexes to -1 to
indicate a free slot at the start, and then do a straightforward
search for the next free slot. And for good measure make sure all
the -1's have disappeared at the end.

v2: Remove spurious whitespace (Jani)
    Fix the post init plane->index assert (Jani/CI)

Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>