tests/intel/xe_exec_system_allocator: Cleanup compute test
Cleanup the test that uses lib/intel_compute and SVM to execute a compute
kernel on arrays allocated on host, in preparation to add more tests.
Remove the result check as this is already done in lib/intel_compute.
Remove output to simplify the test as it did not add value.
Remove the drm_xe_engine_class_instance argument as it was unused.

Cc: Nishit Sharma <nishit.sharma@intel.com>
Cc: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Cc: Matthew Brost <matthew.brost@intel.com>
Reviewed-by: Zbigniew Kempczyński <zbigniew.kempczynski@intel.com>
Signed-off-by: Francois Dugast <francois.dugast@intel.com>