lib/kms: Simplify kmstest_get_vbl_flag()
Simplify kmstest_get_vbl_flag() by observing that:
- '0 << DRM_VBLANK_HIGH_CRTC_SHIFT == 0' so we don't
  have to have special treatment for crtc_offset==0
- the kernel will treat '1 << DRM_VBLANK_HIGH_CRTC_SHIFT'
  and 'DRM_VBLANK_SECONDARY' identically, so the
  crtc_offset==1 special case can also go away.

DRM_VBLANK_HIGH_CRTC_* was added in 2011, so there is no real
point in keeping support for old kernels that don't have it.
And in case someone is crazy enoguh to run igt on a kernel older
than that we can toss in a DRM_CAP_VBLANK_HIGH_CRTC check into
igt_display_require().

Reviewed-by: Jani Nikula <jani.nikula@intel.com>
Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>